Splitted service logic into separate target files
