+++ applications/services/ble/application.fam
@@ -16,7 +16,8 @@ App(
         "service/*.c",
         "service/target/ble_service_target_917.c",
         "service/device_info/*.c",
-        "service/device_events/*.c",
+        "service/device_events/ble_service_device_events.c",
+        "service/device_events/ble_service_device_events_917.c",
         "service/battery/ble_service_battery.c",
         "service/battery/ble_service_battery_917.c",
         "service/uart/*.c",
@@ -42,7 +43,8 @@ App(
         "service/*.c",
         "service/target/ble_service_target_u5.c",
         "service/device_info/*.c",
-        "service/device_events/*.c",
+        "service/device_events/ble_service_device_events.c",
+        "service/device_events/ble_service_device_events_u5.c",
         "service/battery/ble_service_battery.c",
         "service/battery/ble_service_battery_u5.c",
         "service/uart/*.c",
