+++ applications/services/ble/service/device_events/ble_service_device_events.h
@@ -2,4 +2,8 @@
 
 #include "../ble_service_config_types.h"
 
+typedef enum {
+    BleSrvDeviceEventsCharacterIndexFlags,
+} BleSrvDeviceEventsCharacterIndex;
+
 extern const BleServiceDescriptor ble_service_config_device_events;
