+++ applications/services/ble/service/device_events/ble_service_device_events_917.c
@@ -1,0 +1,11 @@
+#include "ble_service_device_events_i.h"
+
+bool ble_service_device_events_init(void* object) {
+    UNUSED(object);
+    return true;
+}
+
+bool ble_service_device_events_run(void* object) {
+    UNUSED(object);
+    return true;
+}
